+++ apps/server/src/mail/mail.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import { SystemModule } from '../system/system.module';
+import { MailController } from './mail.controller';
+import { MailService } from './mail.service';
+
+@Module({
+  imports: [SystemModule],
+  controllers: [MailController],
+  providers: [MailService],
+  exports: [MailService]
+})
+export class MailModule {}